refactor: Update ProxyTestService to return a status object
