+++ alert-common/src/main/java/com/synopsys/integration/alert/common/message/model/ConfigurationTestResult.java
@@ -1,8 +1,14 @@
 package com.synopsys.integration.alert.common.message.model;
 
 public class ConfigurationTestResult {
+    private static final String STATUS_MESSAGE_SUCCESS = "Success";
+
     private final boolean isSuccess;
     private final String statusMessage;
+
+    public static ConfigurationTestResult success() {
+        return new ConfigurationTestResult(true, STATUS_MESSAGE_SUCCESS);
+    }
 
     public static ConfigurationTestResult success(String statusMessage) {
         return new ConfigurationTestResult(true, statusMessage);
