+++ component/src/main/java/com/synopsys/integration/alert/component/settings/proxy/action/SettingsProxyTestAction.java
@@ -7,21 +7,19 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.api.common.model.exception.AlertException;
 import com.synopsys.integration.alert.common.action.ActionResponse;
 import com.synopsys.integration.alert.common.action.ValidationActionResponse;
 import com.synopsys.integration.alert.common.enumeration.ConfigContextEnum;
-import com.synopsys.integration.alert.common.message.model.MessageResult;
-import com.synopsys.integration.alert.common.rest.ProxyTestService;
+import com.synopsys.integration.alert.common.message.model.ConfigurationTestResult;
 import com.synopsys.integration.alert.common.rest.api.ConfigurationTestHelper;
 import com.synopsys.integration.alert.common.rest.api.ConfigurationValidationHelper;
 import com.synopsys.integration.alert.common.rest.model.SettingsProxyModel;
 import com.synopsys.integration.alert.common.rest.model.ValidationResponseModel;
+import com.synopsys.integration.alert.common.rest.proxy.ProxyTestService;
 import com.synopsys.integration.alert.common.security.authorization.AuthorizationManager;
 import com.synopsys.integration.alert.component.settings.descriptor.SettingsDescriptorKey;
 import com.synopsys.integration.alert.component.settings.proxy.database.accessor.SettingsProxyConfigAccessor;
 import com.synopsys.integration.alert.component.settings.proxy.validator.SettingsProxyValidator;
-import com.synopsys.integration.exception.IntegrationException;
 
 @Component
 public class SettingsProxyTestAction {
@@ -46,9 +44,9 @@ public class SettingsProxyTestAction {
         return testHelper.test(validationSupplier, () -> testConfigModelContent(testUrl, requestResource));
     }
 
-    public MessageResult testConfigModelContent(String testUrl, SettingsProxyModel settingsProxyModel) throws AlertException {
+    public ConfigurationTestResult testConfigModelContent(String testUrl, SettingsProxyModel settingsProxyModel) {
         if (StringUtils.isBlank(testUrl)) {
-            throw new AlertException("Could not determine what URL to test the proxy. Target URL was not provided or was blank. Please provide a valid URL to test the configuration.");
+            return ConfigurationTestResult.failure("Could not determine what URL to test the proxy. Target URL was not provided or was blank. Please provide a valid URL to test the configuration.");
         }
 
         if (BooleanUtils.toBoolean(settingsProxyModel.getIsProxyPasswordSet()) && settingsProxyModel.getProxyPassword().isEmpty()) {
@@ -57,11 +55,6 @@ public class SettingsProxyTestAction {
                 .ifPresent(settingsProxyModel::setProxyPassword);
         }
 
-        try {
-            proxyTestService.pingHost(testUrl, settingsProxyModel);
-        } catch (IntegrationException ex) {
-            throw new AlertException(ex.getMessage());
-        }
-        return MessageResult.success();
+        return proxyTestService.pingHost(testUrl, settingsProxyModel);
     }
 }
